refactored ThemeCompiler compiler pass extension
