+++ src/Resources/config/services.xml
@@ -9,6 +9,11 @@
             <tag name="shopware.entity.definition" entity="dne_custom_js_css" />
         </service>
 
+        <service id="Dne\CustomCssJs\Services\FilesystemDecorator" public="true">
+            <argument type="service" id="shopware.filesystem.theme"/>
+            <argument type="service" id="Doctrine\DBAL\Connection"/>
+        </service>
+
         <service id="Dne\CustomCssJs\Api\ThemeCompileController" public="true">
             <argument type="service" id="Shopware\Storefront\Theme\ThemeService"/>
